refactor: move components to presentation layer
