+++ app/components/BarChart/index.tsx
