+++ app/components/BarChart/style.ts
