+++ app/components/BottomSheet/styles.ts
