+++ app/components/DataRangeSelector/CustomContent.tsx
