+++ app/components/DinamicList/index.tsx
