+++ app/components/DinamicList/style.ts
