+++ app/components/Footer/style.ts
