+++ app/components/Illustrations/CatPaw.tsx
