+++ app/components/Illustrations/Coin.tsx
