+++ app/components/Illustrations/Logo.tsx
