+++ app/components/Illustrations/LogoCat.tsx
