+++ app/components/Illustrations/LogoLarge.tsx
