+++ app/components/Illustrations/Waves.tsx
