+++ app/components/ImageCard/style.ts
