+++ app/components/transactions/DinamicTransactionsList/style.ts
