+++ app/components/transactions/StaticTransactionsList/style.ts
